+++ package.json
@@ -4,12 +4,14 @@
   "packageManager": "pnpm@8.6.6+sha256.52f3e7d33884929f1902a7494a616d844b506ed46592e65463ba45e97cc25a1b",
   "scripts": {
     "define-theme": "pnpm --filter @pandabox/define-theme",
+    "knip": "knip",
     "release": "changeset publish"
   },
   "devDependencies": {
     "@arktype/attest": "^0.5.0",
     "@changesets/cli": "^2.27.1",
     "@types/node": "^20.11.4",
+    "knip": "^5.0.0",
     "prettier": "^3.2.2",
     "tsx": "^4.7.0",
     "typescript": "^5.3.3"
