+++ packages/unplugin-panda-macro/package.json
@@ -70,19 +70,16 @@
     "lint": "eslint .",
     "play": "npm -C playground run dev",
     "prepublishOnly": "npm run build",
-    "release": "bumpp && npm publish",
     "start": "esno src/index.ts",
     "test": "vitest"
   },
   "dependencies": {
     "@pandabox/postcss-plugins": "workspace:^",
-    "@pandacss/config": "^0.30.1",
-    "@pandacss/core": "^0.30.1",
-    "@pandacss/extractor": "^0.30.1",
-    "@pandacss/node": "^0.30.1",
-    "@pandacss/parser": "^0.30.1",
-    "@pandacss/shared": "^0.30.1",
-    "@pandacss/types": "^0.30.1",
+    "@pandacss/config": "^0.30.2",
+    "@pandacss/core": "^0.30.2",
+    "@pandacss/extractor": "^0.30.2",
+    "@pandacss/node": "^0.30.2",
+    "@pandacss/shared": "^0.30.2",
     "@rollup/pluginutils": "^5.1.0",
     "magic-string": "^0.30.6",
     "postcss": "^8.4.34",
@@ -90,19 +87,16 @@
     "unplugin": "^1.5.1"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^2.2.2",
     "@nuxt/kit": "^3.8.2",
     "@nuxt/schema": "^3.8.2",
-    "@pandacss/preset-base": "^0.30.1",
-    "@pandacss/preset-panda": "^0.30.1",
+    "@pandacss/preset-base": "^0.30.2",
+    "@pandacss/preset-panda": "^0.30.2",
+    "@pandacss/types": "^0.30.2",
     "@types/node": "^20.10.3",
-    "bumpp": "^9.2.0",
     "chalk": "^5.3.0",
-    "eslint": "^8.55.0",
     "esno": "^4.0.0",
     "fast-glob": "^3.3.2",
     "nodemon": "^3.0.2",
-    "rimraf": "^5.0.5",
     "rollup": "^4.6.1",
     "tsup": "^8.0.1",
     "typescript": "^5.3.2",
