+++ packages/unplugin-panda-macro/playground/minimal.tsx
@@ -1,0 +1,16 @@
+import { css } from './styled-system/css'
+import 'virtual:panda.css'
+
+export const App = () => {
+  return (
+    <>
+      <div
+        className={css({
+          color: 'green',
+        })}
+      >
+        <span>Hello from Panda</span>
+      </div>
+    </>
+  )
+}
