+++ packages/unplugin-panda-macro/playground/package.json
@@ -8,6 +8,7 @@
     "@pandacss/dev": "^0.30.1",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
+    "@vitejs/plugin-react-swc": "^3.6.0",
     "nodemon": "^3.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
