+++ packages/unplugin-panda-macro/src/esbuild.ts
@@ -1,4 +1,4 @@
 import { createEsbuildPlugin } from 'unplugin'
-import { unpluginFactory } from '.'
+import { unpluginFactory } from './plugin/core'
 
 export default createEsbuildPlugin(unpluginFactory)
