+++ packages/unplugin-panda-macro/src/rspack.ts
@@ -1,4 +1,4 @@
 import { createRspackPlugin } from 'unplugin'
-import { unpluginFactory } from '.'
+import { unpluginFactory } from './plugin/core'
 
 export default createRspackPlugin(unpluginFactory)
