+++ packages/unplugin-panda-macro/src/vite.ts
@@ -1,4 +1,4 @@
 import { createVitePlugin } from 'unplugin'
-import { unpluginFactory } from '.'
+import { unpluginFactory } from './plugin/core'
 
 export default createVitePlugin(unpluginFactory)
