+++ website/app/routes/_index.tsx
@@ -27,20 +27,25 @@ export default function Index() {
     <Layout>
       <Center h="100%">
         <Stack p="8" textAlign="center">
-          <styled.h1 textStyle={{ base: '4xl', md: '7xl' }} fontWeight="bold" color="yellow.400">
+          <styled.h1 textStyle={{ base: '4xl', md: '7xl' }} fontWeight="bold" color="blue.400">
             @pandabox
           </styled.h1>
           <styled.h1 textStyle={{ base: '3xl', md: '4xl' }} color={{ base: 'gray.500', _dark: 'gray.300' }}>
             a toolbox for 🐼 Panda CSS
           </styled.h1>
           <Stack gap="1" mt="4" textStyle={{ base: 'sm', md: 'lg' }}>
+            <a className={link} href="https://github.com/astahmer/pandabox/tree/main/packages/unplugin-panda-macro">
+              <LuExternalLink />
+              <span className={linkTitle}>@pandabox/unplugin-panda-macro</span> Make your `styled-system` disappear at
+              build-time by inlining the results as class names.
+            </a>
             <Link className={link} to="/styled2panda">
               <LuLink /> <span className={linkTitle}>styled2panda</span> Easily migrate code from `styled-components` to
               Panda CSS
             </Link>
             <a className={link} href="https://tw2panda-astahmer.vercel.app/" target="_blank">
-              <LuExternalLink /> <span className={linkTitle}>tw2panda</span> : Easily migrate code from tailwind to
-              Panda CSS
+              <LuExternalLink /> <span className={linkTitle}>tw2panda</span> Easily migrate code from tailwind to Panda
+              CSS
             </a>
             <a className={link} href="https://github.com/astahmer/pandabox/tree/main/packages/utils">
               <LuExternalLink />
@@ -58,6 +63,9 @@ export default function Index() {
             >
               <LuExternalLink /> <span className={linkTitle}>defineRecipe</span> Extend, pick, omit and merge config
               Recipes/Slots Recipes to easily compose them together.
+            </a>
+            <a className={link} href="https://www.astahmer.dev" target="_blank">
+              <LuExternalLink /> <span className={linkTitle}>Alex website</span>
             </a>
           </Stack>
         </Stack>
