+++ website/package.json
@@ -16,8 +16,8 @@
     "#styled-system/*": "./styled-system/*"
   },
   "dependencies": {
-    "@ark-ui/anatomy": "^2.1.0",
-    "@ark-ui/react": "^2.0.1",
+    "@ark-ui/anatomy": "^2.2.0",
+    "@ark-ui/react": "^2.1.1",
     "@fontsource/inter": "^5.0.16",
     "@monaco-editor/react": "^4.6.0",
     "@pandabox/codemods": "workspace:^",
@@ -25,28 +25,30 @@
     "@remix-run/node": "^2.6.0",
     "@remix-run/react": "^2.6.0",
     "@remix-run/serve": "^2.6.0",
-    "@xstate/react": "^4.0.3",
-    "isbot": "^4.1.0",
+    "@xstate/react": "^4.1.0",
+    "isbot": "^5.1.0",
     "lz-string": "^1.5.0",
-    "monaco-editor": "^0.45.0",
+    "monaco-editor": "^0.46.0",
     "pastable": "^2.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
-    "react-resizable-panels": "^2.0.3",
-    "ts-morph": "19.0.0",
-    "xstate": "^5.6.1"
+    "react-resizable-panels": "^2.0.9",
+    "ts-morph": "21.0.1",
+    "xstate": "^5.7.1"
   },
   "devDependencies": {
-    "@pandacss/dev": "^0.30.1",
-    "@pandacss/preset-base": "^0.30.1",
-    "@park-ui/panda-preset": "^0.33.0",
+    "@pandabox/unplugin-panda-macro": "workspace:^",
+    "@pandacss/dev": "^0.31.0",
+    "@pandacss/preset-base": "^0.31.0",
+    "@park-ui/panda-preset": "^0.35.1",
     "@remix-run/dev": "^2.6.0",
-    "@types/react": "^18.2.20",
-    "@types/react-dom": "^18.2.7",
+    "@types/react": "^18.2.56",
+    "@types/react-dom": "^18.2.19",
     "dts-bundle-generator": "^9.3.1",
-    "typescript": "^5.1.6",
-    "vite": "^5.0.12"
+    "typescript": "^5.3.3",
+    "vite": "^5.1.3",
+    "vite-plugin-inspect": "^0.8.3"
   },
   "engines": {
     "node": ">=18.0.0"
